# Conflicts:
#	Lab 11/CJ/Project/EEPROM.uvoptx
#	Lab 11/CJ/Project/Objects/EEPROM.htm
#	Lab 11/CJ/Project/Objects/EEPROM_Target 1.dep

+++ Lab 12/Project Folder/ADC.h
@@ -3,7 +3,7 @@
 
 #define VREF 3.3
 
-double Read_ADC(void);
+int Read_ADC(void);
 void adc_init(void);
 
 #endif
